pyproject.toml
[project]
name = "rheinwerk-appliedgenai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.1.0",
    "ipykernel>=6.29.5",
    "langchain-chroma>=0.1.4",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.2.9",
    "langchain>=0.3.7",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.2",
    "swarm",
    "wikipedia>=1.4.0",
    "ag2>=0.3.2",
    "nltk>=3.9.1",
    "langgraph>=0.2.56",
    "langchain-groq>=0.2.1",
    "agentops>=0.3.21",
    "pydantic-ai[logfire]>=0.0.15",
    "nest-asyncio>=1.6.0",
    "langchain-community>=0.3.7",
    "uvicorn>=0.32.1",
    "ragas>=0.2.11",
    "accelerate>=1.3.0",
    "openai-agents>=0.0.7",
    "gradio>=5.43.1",
    "ipywidgets>=8.1.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
]

[tool.uv.sources]
swarm = { git = "https://github.com/openai/swarm.git" }
